# openapi.yaml
openapi: 3.1.0
info:
  title: Dash API
  version: 0.1.0
  description: >
    Single source of truth for the Dash API. The TypeScript client in
    packages/api-client is generated from this file — do not hand-write
    API types in the frontend.
servers:
  - url: /
paths:
...
  /api/sections:
    get:
      operationId: listSections
      summary: Board tree — sections with nested items and urls, ordered
      parameters:
        - name: board
          in: query
          description: Restrict to one board id; omitted returns all sections
          schema:
            type: string
      responses:
        "200":
          description: Ordered board tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: createSection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /api/sections/{id}:
    patch:
      operationId: updateSection
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionPatch"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSection
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Section and its items deleted
  /api/sections/reorder:
    post:
      operationId: reorderSection
      summary: Move a section between two neighbors (fractional position)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reorder"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/systems/ingest:
    post:
      operationId: ingestSystem
      summary: "Agent push endpoint; Authorization: Bearer <system token>"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatSample"
      responses:
        "200":
          description: Sample stored
        "401":
          description: Missing or unknown token
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/auth/setup:
    post:
      operationId: authSetup
      summary: Create the first local admin — only while no users exist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Admin created; session cookie set
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Setup already completed
  /api/auth/login:
    post:
      operationId: authLogin
      summary: Password login — sets a session cookie; rate-limited per IP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Signed in
        "401":
          description: Invalid credentials
        "429":
          description: Too many attempts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Url:
      type: object
      required: [id, url, label]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        label:
          type: string
          description: Tag such as "local" / "external" or a custom label
        position:
          type: number
...
    Item:
      type: object
      required: [id, sectionId, kind, name, position, urls]
      properties:
        id:
          type: string
        sectionId:
          type: string
        kind:
          type: string
          enum: [service, widget]
        name:
          type: string
        icon:
          type: string
          description: "/api/icons/<file> path or remote URL; empty = letter tile"
        position:
          type: number
        config:
          type: object
          nullable: true
          additionalProperties: true
          description: Widget config (Phase 2); null for services
        urls:
          type: array
          items:
            $ref: "#/components/schemas/Url"
...
    Section:
      type: object
      required: [id, name, position, collapsed, boardId, items]
      properties:
        id:
          type: string
        name:
          type: string
        position:
          type: number
        collapsed:
          type: boolean
        boardId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
...
    SectionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        boardId:
          type: string
    SectionPatch:
      type: object
      properties:
        name:
          type: string
        collapsed:
          type: boolean
    Reorder:
      type: object
      required: [id]
      properties:
        id:
          type: string
        beforeId:
          type: string
          description: Neighbor now before the moved entity; null if first
        afterId:
          type: string
          description: Neighbor now after the moved entity; null if last
...
    StatSample:
      type: object
      description: One agent push; missing fields mean the platform can't provide them
      properties:
        host: { type: string }
        os: { type: string }
        arch: { type: string }
        cpuModel: { type: string }
        cores: { type: integer }
        intervalS: { type: integer }
        uptimeS: { type: number }
        cpu:
          type: number
          description: CPU usage percent 0-100
        memTotal: { type: number, description: bytes }
        memUsed: { type: number, description: bytes }
        swapTotal: { type: number, description: bytes }
        swapUsed: { type: number, description: bytes }
        diskTotal: { type: number, description: root filesystem bytes }
        diskUsed: { type: number, description: bytes }
        netRx: { type: number, description: bytes per second }
        netTx: { type: number, description: bytes per second }
        load1: { type: number }
        load5: { type: number }
        load15: { type: number }
        temps:
          type: object
          additionalProperties: { type: number }
          description: Sensor label -> celsius
        containers:
          type: array
          items:
            $ref: "#/components/schemas/ContainerStat"
        smart:
          type: array
          items:
            $ref: "#/components/schemas/SmartDisk"
        zfs:
          type: array
          items:
            $ref: "#/components/schemas/ZFSPool"
        gpu:
          type: array
          items:
            $ref: "#/components/schemas/GPUStat"
    ContainerStat:
      type: object
      properties:
        name: { type: string }
        state: { type: string }
        cpu: { type: number, description: "percent of one core, may exceed 100" }
        memUsed: { type: number, description: bytes }
        memLimit: { type: number, description: bytes }
    SmartDisk:
      type: object
      properties:
        device: { type: string }
        model: { type: string }
        passed: { type: boolean, nullable: true }
        tempC: { type: number }
    ZFSPool:
      type: object
      properties:
        name: { type: string }
        health: { type: string }
        size: { type: number, description: bytes }
        free: { type: number, description: bytes }
    GPUStat:
      type: object
      properties:
        name: { type: string }
        tempC: { type: number }
        utilPct: { type: number }
        memUsed: { type: number, description: bytes }
        memTotal: { type: number, description: bytes }
...